update hip-sycl interop samples to use the new API

+++ samples/hip_sycl_interop_no_buffers/hip_sycl_interop.h
@@ -3,7 +3,7 @@
 
 extern "C" {
   // Run GEMM test via oneMKL
-  int oneMKLGemmTest(unsigned long* nativeHandlers, float* A, float* B, float* C, int M, int N, int K,
+  int oneMKLGemmTest(uintptr_t* nativeHandlers, float* A, float* B, float* C, int M, int N, int K,
 		     int ldA, int ldB, int ldC, float alpha, float beta);
 }
